Refactored code out of main and into seperate lib files.
