+++ lib/jquery.js
