+++ lib/noCache.js
@@ -1,0 +1,7 @@
+class NoCache {
+    get(url, cb) { cb(null)}
+    getHeaders(url, cb) {cb(null)}
+    set(url, headers, body) {}
+}
+
+module.exports = NoCache;
